Moving retrieve with constraints test into retrieve test group
